+++ include/flecs.h
@@ -90,16 +90,16 @@ typedef struct ecs_world_info_t {
     ecs_entity_t min_id;              /**< First allowed entity id */
     ecs_entity_t max_id;              /**< Last allowed entity id */
 
-    float delta_time_raw;      /**< Raw delta time (no time scaling) */
-    float delta_time;          /**< Time passed to or computed by ecs_progress */
-    float time_scale;          /**< Time scale applied to delta_time */
-    float target_fps;          /**< Target fps */
-    float frame_time_total;    /**< Total time spent processing a frame */
-    float system_time_total;   /**< Total time spent in systems */
-    float merge_time_total;    /**< Total time spent in merges */
-    float world_time_total;    /**< Time elapsed in simulation */
-    float world_time_total_raw; /**< Time elapsed in simulation (no scaling) */
-    float sleep_err;           /**< Measured sleep error */
+    double delta_time_raw;      /**< Raw delta time (no time scaling) */
+    double delta_time;          /**< Time passed to or computed by ecs_progress */
+    double time_scale;          /**< Time scale applied to delta_time */
+    double target_fps;          /**< Target fps */
+    double frame_time_total;    /**< Total time spent processing a frame */
+    double system_time_total;   /**< Total time spent in systems */
+    double merge_time_total;    /**< Total time spent in merges */
+    double world_time_total;    /**< Time elapsed in simulation */
+    double world_time_total_raw; /**< Time elapsed in simulation (no scaling) */
+    double sleep_err;           /**< Measured sleep error */
     
     int32_t frame_count_total;  /**< Total number of frames */
     int32_t merge_count_total;  /**< Total number of merges */
@@ -2627,9 +2627,9 @@ int32_t ecs_table_component_index(
 
 /** Begin frame. */
 FLECS_EXPORT
-float ecs_frame_begin(
+double ecs_frame_begin(
     ecs_world_t *world,
-    float delta_time);
+    double delta_time);
 
 /** End frame. */
 FLECS_EXPORT
